MAS-311 split out mocked server setup
