+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppstier/integration/MockedEndpointsTestBase.kt
@@ -1,0 +1,29 @@
+package uk.gov.justice.digital.hmpps.hmppstier.integration
+
+import org.junit.jupiter.api.AfterAll
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.BeforeAll
+import org.mockserver.integration.ClientAndServer
+
+abstract class MockedEndpointsTestBase: IntegrationTestBase() {
+  lateinit var mockCommunityApiServer: ClientAndServer
+  lateinit var mockAssessmentApiServer: ClientAndServer
+
+  @BeforeAll
+  fun setupMockServer() {
+    mockCommunityApiServer = ClientAndServer.startClientAndServer(8081)
+    mockAssessmentApiServer = ClientAndServer.startClientAndServer(8082)
+  }
+
+  @AfterEach
+  fun reset() {
+    mockCommunityApiServer.reset()
+    mockAssessmentApiServer.reset()
+  }
+
+  @AfterAll
+  fun tearDownServer() {
+    mockCommunityApiServer.stop()
+    mockAssessmentApiServer.stop()
+  }
+}
